+++ CHANGELOG.md
@@ -81,6 +81,8 @@ An EtherCAT master written in Rust.
   `SlaveGroup::<PreOp>::request_into_op` to request all SubDevices in a group transition to OP, but
   does not wait for them to transition. Also add `SlaveGroup::<Op>::all_op` to check if all
   SubDevices in the group have reached OP state.
+- [#TODO](https://github.com/ethercrab-rs/ethercrab/pull/#TODO) Add `Error::DistributedClock(_)` and
+  `DistributedClockError` error variant and type to communicate DC errors.
 
 ### Fixed